productpage decode readme
