+++ src/App/App.scss
@@ -1,2 +1,3 @@
 .app {
+  height: 100%;
 }
