+++ src/App/pages/OrgsPage/OrgsPage.module.scss
@@ -1,4 +1,6 @@
 .orgs {
+  margin-top: var(--space-xxl);
+  padding-bottom: var(--space-xxl);
   &__header {
     text-align: center;
   }
