+++ src/App/pages/OrgsPage/components/Pagination/Pagination.module.scss
@@ -1,3 +1,16 @@
+// @import "../../../../../styles/mixins.scss";
+@import "styles/mixins.scss";
+
 .pagination {
   margin-top: var(--margin-xxl);
+  @include flexbox(center, center, 50px);
+  .pagination__list {
+    @include flexbox(center, center, 20px);
+    .pagination__item {
+    }
+  }
+  .pagination__right {
+  }
+  .pagination__left {
+  }
 }
