+++ src/App/pages/OrgsPage/components/Pagination/Pagination.tsx
@@ -8,13 +8,15 @@ const Pagination: FC<IPaginationProps> = () => {
   const num = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 114, 15];
   return (
     <div className={css.pagination}>
-      <ul className={css.pages}>
+      <button className={css.pagination__left}>left</button>
+      <ul className={css.pagination__list}>
         {num.map((el) => (
-          <li key={el} className={css.page}>
+          <li key={el} className={css.pagination__item}>
             {el}
           </li>
         ))}
       </ul>
+      <button className={css.pagination__right}>right</button>
     </div>
   );
 };
