+++ src/components/Card/Card.module.scss
@@ -1,29 +1,26 @@
+@import "styles/mixins.scss";
+
 .card {
   max-width: 360px;
-  // max-height: 598px;
-  max-height: 622px;
+  // height: 506px;
+  // height: 622px;
 
   background-color: var(--card-bg);
   &:hover {
     box-shadow: 0px 4px 10px 0px rgba(0, 0, 0, 0.25);
   }
   .card__img {
-    max-height: 360px;
     width: 100%;
+    max-height: 360px;
+
     background-color: var(--card-image-bg);
-    display: block;
   }
   .card__info {
     padding: var(--space-xl) var(--space-xs);
-    .card__header {
-      .card__caption {
-        margin-bottom: var(--space-xxs);
-      }
-    }
+
     .card__footer {
-      display: flex;
-      align-items: center;
-      justify-content: space-between;
+      @include flexbox(center, space-between);
+      margin-top: 20px;
     }
   }
 }
