+++ src/components/Input/Input.module.scss
@@ -1,7 +1,7 @@
+@import "styles/mixins.scss";
+
 .inputWrapper {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
+  @include flexbox(center, space-between);
 
   width: 100%;
   height: 52px;
