+++ src/components/Text/Text.module.scss
@@ -44,3 +44,7 @@
   font-size: 14px;
   line-height: 18px;
 }
+.p-12 {
+  font-size: 12px;
+  line-height: 14px;
+}
