+++ src/components/Text/Text.tsx
@@ -13,7 +13,7 @@ export type TextProps = {
   /** Дополнительный класс */
   className?: string;
   /** Стиль отображения */
-  view?: 'title' | 'button' | 'p-20' | 'p-18' | 'p-16' | 'p-14';
+  view?: 'title' | 'button' | 'p-20' | 'p-18' | 'p-16' | 'p-14' | 'p-12';
   /** Html-тег */
   tag?: 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6' | 'div' | 'p' | 'span' | 'li';
   /** Начертание шрифта */
@@ -39,7 +39,7 @@ const Text: React.FC<TextProps> = ({
   mb,
   ml,
   mr,
-  m,
+  m = 0,
   stylesProps,
 }) => {
   const viewCls = view && css[view];
@@ -51,11 +51,11 @@ const Text: React.FC<TextProps> = ({
     fontWeight: weight,
     lineClamp: maxLines,
     WebkitLineClamp: maxLines,
+    margin: m,
     marginTop: mt,
-    marginBottom: mb,
     marginLeft: ml,
     marginRight: mr,
-    margin: m,
+    marginBottom: mb,
     ...stylesProps,
   };
