+++ src/main.tsx
@@ -3,7 +3,7 @@ import './styles/index.scss';
 import { StrictMode } from 'react';
 import ReactDOM from 'react-dom/client';
 
-import App from 'App';
+import App from 'App/App';
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <StrictMode>
