+++ src/styles/index.scss
@@ -6,18 +6,31 @@
 
 html,
 body,
-#root {
-  font-family: Roboto;
+#root,
+.app {
   height: 100%;
+  // min-height: 100%;
 }
 
 body {
-  width: 1440px;
+  font-family: Roboto;
+  font-weight: 400;
+  font-size: 16px;
+  // min-height: 100vh;
+}
+
+#root {
+  max-width: 1440px;
   margin: 0 auto;
-  background: #f6f8fa;
+}
+.app {
+  background-color: #ffa1bd;
+  // background-color: #f6f8fa;
 }
 
 .container {
+  margin: 0 auto;
   padding: 0 100px;
+  overflow: hidden;
+  // min-height: 100%;
 }
-// --main-gb: #F6F8FA;
