+++ src/styles/mixins.scss
@@ -1,0 +1,6 @@
+@mixin flexbox($align-items, $justify-content, $gap: 0) {
+  display: flex;
+  align-items: $align-items;
+  justify-content: $justify-content;
+  gap: $gap;
+}
